--- src/pages/Dashboard/PaymentCredentials/api.js
@@ -1,7 +1,0 @@
-import axios from 'axios';
-
-const instance = axios.create({
-  baseURL: process.env.REACT_APP_API_BASE_URL
-});
-
-export default instance;
